pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docintel"
version = "0.1.0"
description = "Windows-first file governance platform with SQLite-backed inventory, planning, validation, and safe execution."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openpyxl>=3.1.5",
  "pypdf>=4.3.1",
  "python-docx>=1.1.2",
]

[project.optional-dependencies]
gui = ["PySide6>=6.7,<7"]
dev = [
  "mypy>=1.11",
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.11",
]

[tool.setuptools.packages.find]
include = ["docintel*", "config*", "db*", "extractors*", "logs*", "scanner*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
